Rename test_config -> test-config
